# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GrailBabe API - collectibles portfolio tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      parameters:
        - { name: category, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TradePost" }
    post:
      operationId: createTrade
      tags: [trades]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TradePostInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TradePost" }
  /trades/mine:
    get:
      operationId: listMyTrades
      tags: [trades]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, open, pending, completed, cancelled]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MyTrade" }
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TradePost" }
    delete:
      operationId: deleteTrade
      tags: [trades]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /trades/{id}/offer:
    post:
      operationId: makeTradeOffer
      tags: [trades]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MyTrade" }
  /trades/{id}/confirm-complete:
    post:
      operationId: confirmTradeComplete
      tags: [trades]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MyTrade" }
  /trades/{id}/cancel:
    post:
      operationId: cancelTrade
      tags: [trades]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MyTrade" }
  /trades/{id}/review:
    post:
      operationId: leaveTradeReview
      tags: [trades]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveReviewInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MyTrade" }
  /users/{userId}/reviews:
    get:
      operationId: listUserReviews
      tags: [profiles]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserReview" }
...
components:
  schemas:
...
    TradePost:
      type: object
      required: [id, userId, userName, title, category, condition, kind, createdAt]
      properties:
        id: { type: integer }
        userId: { type: string }
        userName: { type: string }
        userAvatar: { type: ["string", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        category: { type: string }
        condition: { type: string }
        askingPrice: { type: ["number", "null"] }
        photos:
          type: array
          items: { type: string }
        kind: { type: string, enum: [trade, sell, buy] }
        wantedItems:
          type: array
          items: { type: string }
        vaultItemId: { type: ["integer", "null"] }
        status:
          type: string
          enum: [open, pending, completed, cancelled, closed]
        createdAt: { type: string, format: date-time }
    MyTradeOtherParty:
      type: object
      required: [id, screenname, displayName, initials]
      properties:
        id: { type: string }
        screenname: { type: string }
        displayName: { type: string }
        initials: { type: string }
        avatarUrl: { type: ["string", "null"] }
    MyTradeReviewSummary:
      type: object
      required: [rating, createdAt]
      properties:
        rating: { type: integer }
        comment: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    MyTrade:
      type: object
      required:
        - id
        - role
        - status
        - title
        - category
        - condition
        - kind
        - wantedItems
        - photos
        - myConfirmed
        - theirConfirmed
        - postedAt
      properties:
        id: { type: integer }
        role: { type: string, enum: [poster, requester] }
        status:
          type: string
          enum: [open, pending, completed, cancelled, closed]
        title: { type: string }
        description: { type: ["string", "null"] }
        category: { type: string }
        condition: { type: string }
        askingPrice: { type: ["number", "null"] }
        kind: { type: string, enum: [trade, sell, buy] }
        wantedItems:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string }
        vaultItemId: { type: ["integer", "null"] }
        otherParty:
          oneOf:
            - { $ref: "#/components/schemas/MyTradeOtherParty" }
            - { type: "null" }
        myConfirmed: { type: boolean }
        theirConfirmed: { type: boolean }
        postedAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        myReview:
          oneOf:
            - { $ref: "#/components/schemas/MyTradeReviewSummary" }
            - { type: "null" }
        theirReview:
          oneOf:
            - { $ref: "#/components/schemas/MyTradeReviewSummary" }
            - { type: "null" }
    LeaveReviewInput:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: ["string", "null"], maxLength: 300 }
    UserReview:
      type: object
      required: [id, tradeId, rating, reviewerId, reviewerScreenname, reviewerInitials, createdAt]
      properties:
        id: { type: integer }
        tradeId: { type: integer }
        rating: { type: integer }
        comment: { type: ["string", "null"] }
        reviewerId: { type: string }
        reviewerScreenname: { type: string }
        reviewerInitials: { type: string }
        createdAt: { type: string, format: date-time }
    TradePostInput:
      type: object
      required: [title, category, condition, kind]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        category: { type: string, minLength: 1 }
        condition: { type: string }
        askingPrice: { type: number }
        photos:
          type: array
          items: { type: string }
        kind: { type: string, enum: [trade, sell, buy] }
        wantedItems:
          type: array
          items: { type: string }
        vaultItemId: { type: integer }
...